--- Tests/Tests.Playground/CurrentTests.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Linq;
-using LinqToDB.Mapping;
-using NUnit.Framework;
-
-namespace Tests.Playground
-{
-	[TestFixture]
-    public class CurrentTests : TestBase
-    {
-		[Table]
-	    class SampleClass
-	    {
-		    [Column] public int Id    { get; set; }
-		    [Column] public int Value { get; set; }
-	    }
-
-
-		[Test, Combinatorial]
-	    public void SampleSelectTest([DataSources] string context)
-	    {
-		    using (var db = GetDataContext(context))
-		    using (var table = db.CreateLocalTable<SampleClass>())
-		    {
-			    var result = table.ToArray();
-		    }
-	    }
-
-    }
-}
