+++ linq2db.playground.sln
@@ -23,20 +23,20 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|x86.ActiveCfg = Debug|x86
-		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|x86.Build.0 = Debug|x86
+		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Debug|x86.Build.0 = Debug|Any CPU
 		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|x86.ActiveCfg = Release|x86
-		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|x86.Build.0 = Release|x86
+		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|x86.ActiveCfg = Release|Any CPU
+		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7}.Release|x86.Build.0 = Release|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|x86.ActiveCfg = Travis|x86
-		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|x86.Build.0 = Debug|x86
+		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|x86.Build.0 = Debug|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|x86.ActiveCfg = Release|x86
-		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|x86.Build.0 = Release|x86
+		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|x86.ActiveCfg = Release|Any CPU
+		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|x86.Build.0 = Release|Any CPU
 		{4E3D4705-F037-4536-A45D-7AF52BB017DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4E3D4705-F037-4536-A45D-7AF52BB017DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4E3D4705-F037-4536-A45D-7AF52BB017DC}.Debug|x86.ActiveCfg = Debug|x86
@@ -47,13 +47,12 @@ Global
 		{4E3D4705-F037-4536-A45D-7AF52BB017DC}.Release|x86.Build.0 = Release|x86
 		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|x86.ActiveCfg = Debug|x86
-		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|x86.Build.0 = Debug|x86
-		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|x86.Build.2 = Travis|x86
+		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Debug|x86.Build.0 = Debug|Any CPU
 		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|x86.ActiveCfg = Release|x86
-		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|x86.Build.0 = Release|x86
+		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|x86.ActiveCfg = Release|Any CPU
+		{0C6A618A-CC21-4FD0-8456-AA2BCF56AAE1}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
